redatoraçao com criacao de components para a home
