+++ src/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
-import SearchBar from './Pages/SearchBar';
+import Home from './Pages/Home';
 import './App.css';
 
 function App() {
   return (
     <BrowserRouter>
       <Switch>
-        <Route path="/" exact component={ SearchBar } />
+        <Route path="/" exact component={ Home } />
       </Switch>
     </BrowserRouter>
   );
