+++ src/Components/ProductCard.js
@@ -1,14 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import NotFound from './NotFound';
 
 export default class ProductCard extends Component {
   render() {
     const { product } = this.props;
     const { title, thumbnail, price } = product;
-    if (!product) {
-      return <NotFound />;
-    }
     return (
       <section data-testid="product">
         <h4>{ title }</h4>
